Add encryptfile system, Remove printer example
Remove printer example for walk within `cmd/encrypt.go` and add the
encryptfile code. Update order of directory check and `fileaction.Do`.
This is because it would run the file action on the directory, we do not
want this and we just want it to `continue`.

Add encryptfile system. This was done by using AES with CTR. It is done
by reading block by block from the file. This was choosen so we can
read **HUGE** files without worrying about memory. This also gives us
the major benefit of concurrently reading files without using insane
amounts of memory. after it writes the output to `<filePath> + ".enc"`, it
renames it to just `<filePath>`. This was picked as there is no real
way(without alot of work) to read and write to the same file at the same
time without reading everything into memory.